+++ torchmetrics/functional/audio/sdr.py
@@ -16,7 +16,14 @@ from typing import Tuple
 
 import numpy as np
 import torch
-from mir_eval.separation import bss_eval_sources
+
+from torchmetrics.utilities.imports import _MIR_EVAL_AVAILABLE
+
+if _MIR_EVAL_AVAILABLE:
+    from mir_eval.separation import bss_eval_sources
+else:
+    bss_eval_sources = None
+
 from torch import Tensor
 
 from torchmetrics.utilities.checks import _check_same_shape
@@ -25,30 +32,49 @@ from torchmetrics.utilities.checks import _check_same_shape
 def _sdr_sir_sar(
     preds: Tensor, target: Tensor, compute_permutation: bool = False, keep_same_device: bool = False
 ) -> Tuple[Tensor, Tensor, Tensor]:
-    r"""sdr_sir_sar evaluates the SDR, SIR, SAR metrics of preds and target. sdr_sir_sar is a wrapper for the mir_eval.separation.bss_eval_sources function.
+    r"""_sdr_sir_sar evaluates the SDR, SIR, SAR metrics of preds and target. _sdr_sir_sar is a wrapper for the 
+    mir_eval.separation.bss_eval_sources function.
 
     Args:
         preds:
-            shape ``[..., spk, time]``
+            shape ``[..., time]`` if compute_permutation is False, else ``[..., spk, time]``
         target:
-            shape ``[..., spk, time]``
+            shape ``[..., time]`` if compute_permutation is False, else ``[..., spk, time]``
         compute_permutation:
-            whether to compute the metrics permutation invariantly. By default, it is False for we can use PIT to compute the permutation in a better way and in the sense of any metrics.
+            whether to compute the metrics permutation invariantly. By default, it is False
+            for we can use PIT to compute the permutation in a better way and in the sense of any metrics.
         keep_same_device:
             whether to move the metric value to the device of preds
 
+    Raises:
+        ValueError:
+            If ``mir_eval`` package is not installed, or 1D input is given when compute_permutation is True
+
     Returns:
-        sdr value of shape [..., spk]
-        sir value of shape [..., spk]
-        sar value of shape [..., spk]
+        sdr value of shape ``[...]`` if compute_permutation is False, else ``[..., spk]``
+        sir value of shape ``[...]`` if compute_permutation is False, else ``[..., spk]``
+        sar value of shape ``[...]`` if compute_permutation is False, else ``[..., spk]``
 
     """
+    if not _MIR_EVAL_AVAILABLE:
+        raise ValueError(
+            "SDR metric requires that mir_eval is installed."
+            "Either install as `pip install torchmetrics[audio]` or `pip install mir_eval`"
+        )
+
     _check_same_shape(preds, target)
 
     if preds.ndim == 1:
-        warnings.warn("preds and target should be of the shape [..., spk, time], but 1d tensor detected")
+        if compute_permutation:
+            raise ValueError(
+                "SDR metric requires preds and target to be of shape [..., spk, time]"
+                " if compute_permutation is True, but 1D Tensor is given."
+            )
+
         sdr_val_np, sir_val_np, sar_val_np, perm = bss_eval_sources(
-            target.detach().cpu().numpy()[None, ...], preds.detach().cpu().numpy()[None, ...], compute_permutation
+            target.detach().cpu().numpy()[None, ...],
+            preds.detach().cpu().numpy()[None, ...],
+            compute_permutation,
         )
         sdr_val = torch.tensor(sdr_val_np[0])
         sir_val = torch.tensor(sir_val_np[0])
@@ -89,28 +115,35 @@ def _sdr_sir_sar(
 
 
 def sdr(preds: Tensor, target: Tensor, compute_permutation: bool = False, keep_same_device: bool = False) -> Tensor:
-    r"""sdr evaluates the Signal to Distortion Ratio (SDR) metric of preds and target. sdr is a wrapper for the mir_eval.separation.bss_eval_sources function.
+    r"""sdr evaluates the Signal to Distortion Ratio (SDR) metric of preds and target. sdr is a wrapper for the 
+    mir_eval.separation.bss_eval_sources function.
 
     Args:
         preds:
-            shape ``[..., spk, time]``
+            shape ``[..., time]`` if compute_permutation is False, else ``[..., spk, time]``
         target:
-            shape ``[..., spk, time]``
+            shape ``[..., time]`` if compute_permutation is False, else ``[..., spk, time]``
         compute_permutation:
-            whether to compute the metrics permutation invariantly. By default, it is False for we can use PIT to compute the permutation in a better way and in the sense of any metrics.
+            whether to compute the metrics permutation invariantly. By default, it is False
+            for we can use PIT to compute the permutation in a better way and in the sense of any metrics.
         keep_same_device:
             whether to move the metric value to the device of preds
 
+    Raises:
+        ValueError:
+            If ``mir_eval`` package is not installed, or 1D input is given when compute_permutation is True
+
     Returns:
-        sdr value of shape [..., spk]
+        sdr value of shape ``[...]`` if compute_permutation is False, else ``[..., spk]``
 
     Example:
         >>> from torchmetrics.functional.audio import sdr
         >>> import torch
-        >>> preds = torch.randn(2, 8000)
-        >>> target = torch.randn(2, 8000)
-        >>> sdr_val = sdr(preds, target)
+        >>> g = torch.manual_seed(1)
+        >>> preds = torch.randn(8000)
+        >>> target = torch.randn(8000)
+        >>> sdr(preds, target).float()
+        tensor(-12.0589)
     """
 
     return _sdr_sir_sar(preds, target, compute_permutation, keep_same_device)[0]
-